Adding features: Crud Products, auth, contact

+++ src/app/products/features/product-list/product-list.component.ts
@@ -1,44 +1,34 @@
-import { Component, OnInit, inject, signal } from "@angular/core";
-import { Product } from "app/products/data-access/product.model";
-import { ProductsService } from "app/products/data-access/products.service";
-import { ProductFormComponent } from "app/products/ui/product-form/product-form.component";
-import { ButtonModule } from "primeng/button";
-import { CardModule } from "primeng/card";
-import { DataViewModule } from 'primeng/dataview';
-import { DialogModule } from 'primeng/dialog';
+import {Component, OnInit, inject, signal} from "@angular/core";
+import {Product} from "app/products/data-access/product.model";
+import {ProductsService} from "app/products/data-access/products.service";
+import {ProductFormComponent} from "app/products/ui/product-form/product-form.component";
+import {ButtonModule} from "primeng/button";
+import {CardModule} from "primeng/card";
+import {DataViewModule} from 'primeng/dataview';
+import {DialogModule} from 'primeng/dialog';
+import {throwError} from "rxjs";
+import {CurrencyPipe} from "@angular/common";
+import {CartService} from "../../../services/cart.service";
+import {AuthService} from "../../../services/auth.service";
 
-const emptyProduct: Product = {
-  id: 0,
-  code: "",
-  name: "",
-  description: "",
-  image: "",
-  category: "",
-  price: 0,
-  quantity: 0,
-  internalReference: "",
-  shellId: 0,
-  inventoryStatus: "INSTOCK",
-  rating: 0,
-  createdAt: 0,
-  updatedAt: 0,
-};
 
 @Component({
   selector: "app-product-list",
   templateUrl: "./product-list.component.html",
   styleUrls: ["./product-list.component.scss"],
   standalone: true,
-  imports: [DataViewModule, CardModule, ButtonModule, DialogModule, ProductFormComponent],
+  imports: [DataViewModule, CardModule, ButtonModule, DialogModule, ProductFormComponent, CurrencyPipe],
 })
 export class ProductListComponent implements OnInit {
   private readonly productsService = inject(ProductsService);
+  private readonly cartService = inject(CartService);
+  auth = inject(AuthService);
 
   public readonly products = this.productsService.products;
 
   public isDialogVisible = false;
   public isCreation = false;
-  public readonly editedProduct = signal<Product>(emptyProduct);
+  public readonly editedProduct = signal<Product>(new Product());
 
   ngOnInit() {
     this.productsService.get().subscribe();
@@ -47,7 +37,7 @@ export class ProductListComponent implements OnInit {
   public onCreate() {
     this.isCreation = true;
     this.isDialogVisible = true;
-    this.editedProduct.set(emptyProduct);
+    this.editedProduct.set(new Product());
   }
 
   public onUpdate(product: Product) {
@@ -57,7 +47,8 @@ export class ProductListComponent implements OnInit {
   }
 
   public onDelete(product: Product) {
-    this.productsService.delete(product.id).subscribe();
+    if (product.id)
+      this.productsService.delete(product.id).subscribe();
   }
 
   public onSave(product: Product) {
@@ -76,4 +67,11 @@ export class ProductListComponent implements OnInit {
   private closeDialog() {
     this.isDialogVisible = false;
   }
+
+  addToCart(product: Product): void {
+    if (product.inventoryStatus !== 'OUTOFSTOCK') {
+      this.cartService.addToCart(product);
+    }
+  }
+
 }
